refactor(web): move `position.util` inside `position/`
