+++ packages/web/src/common/utils/position/position.util.ts
@@ -13,8 +13,8 @@ import {
   EVENT_PADDING_RIGHT,
   GRID_MARGIN_LEFT,
 } from "@web/views/Calendar/layout.constants";
-import { ACCEPTED_TIMES } from "../constants/web.constants";
-import { Schema_GridEvent } from "../types/web.event.types";
+import { ACCEPTED_TIMES } from "../../constants/web.constants";
+import { Schema_GridEvent } from "../../types/web.event.types";
 
 dayjs.extend(dayOfYear);
 dayjs.extend(weekPlugin);
